Protecting resources with authorization.

+++ backend/routes/posts.js
@@ -69,8 +69,13 @@ router.put(
     imagePath: imagePath
   });
   console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: "Update Successful!"});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0){
+      res.status(200).json({message: "Update Successful!"});
+    } else {
+      res.status(401).json({message: "Not Authorized!"});
+    }
+
   });
 });
 
@@ -110,12 +115,12 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
-  }).catch(error =>  {
-    console.log(error); // Check if there's any error in the server logs
-    res.status(500).json({ error: "An error occurred while deleting the post." });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0){
+      res.status(200).json({message: "Deletion Successful!"});
+    } else {
+      res.status(401).json({message: "Not Authorized!"});
+    }
   });
 });
 module.exports = router;
